feat: enable renderOnlyVisible with panelsPerView
