--- src/renderer/RenderingStrategy/index.ts
@@ -1,11 +1,0 @@
-import RenderingStrategy from "./RenderingStrategy";
-import RawRenderingStrategy from "./RawRenderingStrategy";
-import VisibleRenderingStrategy from "./VisibleRenderingStrategy";
-import GridRenderingStrategy from "./GridRenderingStrategy";
-
-export {
-  RenderingStrategy,
-  RawRenderingStrategy,
-  VisibleRenderingStrategy,
-  GridRenderingStrategy
-};
